% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIndent.R
\name{assignIndent}
\alias{assignIndent}
\title{assignIndent}
\usage{
assignIndent()
}
\description{
Call this function as an addin to insert an assignment (<-) and a space at the cursor position with an extra indent level.
}
